쿠키-세션 구현

+++ Quest10/client/public/js/storage.mjs
@@ -8,20 +8,33 @@ export default class NotepadStorage {
 
   async getFileList() {
     try {
-      const res = await fetch(dummy.fileUrl);
+      const res = await this.#customFetch();
+      if (res.status === 401) {
+        throw { msg: '권한이 없습니다. \n다시 로그인 해주세요.' };
+      }
       const { fileList } = await res.json();
 
       this.fileData = fileList.reduce((acc, cur) => ({ ...acc, [cur]: null }), {});
       return fileList;
     } catch (err) {
-      alert('파일 목록 로드 에러');
+      if (err.msg) {
+        alert(err.msg);
+        location.href = '/';
+      } else {
+        alert('파일 목록 로드 에러');
+      }
     }
   }
 
   async createFile(newFileName) {
     try {
-      const res = await fetch(dummy.fileUrl, {
+      if (!this.#validation(newFileName)) {
+        throw { msg: '잘못된 파일명입니다' };
+      }
+
+      const res = await this.#customFetch('', {
         method: 'POST',
+        credentials: 'include',
         headers: {
           'Content-type': 'application/json; charset=UTF-8',
         },
@@ -35,17 +48,23 @@ export default class NotepadStorage {
       this.fileData[newFileName] = '';
     } catch (err) {
       err.msg ? alert(err.msg) : alert('파일 생성 에러');
-      throw err;
     }
   }
 
   async getFile(fileName) {
     try {
+      if (!this.#validation(fileName)) {
+        throw { msg: '잘못된 파일명입니다' };
+      }
+
       if (this.fileData[fileName] !== null) {
         return this.fileData[fileName];
       }
 
-      const res = await fetch(dummy.fileUrl + fileName);
+      const res = await this.#customFetch(fileName, {
+        credentials: 'include',
+      });
+
       if (res.status === 404) {
         throw { msg: await res.json() };
       }
@@ -55,14 +74,18 @@ export default class NotepadStorage {
       return data;
     } catch (err) {
       err.msg ? alert(err.msg) : alert('파일 로드 에러');
-      throw err;
     }
   }
 
   async saveFile(fileName, text) {
     try {
-      const res = await fetch(dummy.fileUrl + fileName, {
+      if (!this.#validation(fileName)) {
+        throw { msg: '잘못된 파일명입니다' };
+      }
+
+      const res = await this.#customFetch(fileName, {
         method: 'PUT',
+        credentials: 'include',
         headers: {
           'Content-type': 'application/json; charset=UTF-8',
         },
@@ -79,14 +102,18 @@ export default class NotepadStorage {
       alert('파일 저장 완료');
     } catch (err) {
       err.msg ? alert(err.msg) : alert('파일 저장 실패');
-      throw err;
     }
   }
 
   async saveAsFile(oldFileName, newFileName, text) {
     try {
-      const res = await fetch(dummy.fileUrl + oldFileName, {
+      if (!this.#validation(oldFileName) && !this.#validation(newFileName)) {
+        throw { msg: '잘못된 파일명입니다' };
+      }
+
+      const res = await this.#customFetch(oldFileName, {
         method: 'POST',
+        credentials: 'include',
         headers: {
           'Content-type': 'application/json; charset=UTF-8',
         },
@@ -106,14 +133,18 @@ export default class NotepadStorage {
       alert('다른 이름으로 저장 완료');
     } catch (err) {
       err.msg ? alert(err.msg) : alert('저장 실패');
-      throw err;
     }
   }
 
   async removeFile(fileName) {
     try {
-      const res = await fetch(dummy.fileUrl + fileName, {
+      if (!this.#validation(fileName)) {
+        throw { msg: '잘못된 파일명입니다' };
+      }
+
+      const res = await this.#customFetch(fileName, {
         method: 'DELETE',
+        credentials: 'include',
       });
 
       if (res.status !== 200) {
@@ -126,4 +157,19 @@ export default class NotepadStorage {
       err.msg ? alert(err.msg) : alert('파일 삭제 실패');
     }
   }
+
+  #validation(fileName) {
+    return ['/', '.'].some((x) => fileName.includes(x)) ? false : true;
+  }
+
+  async #customFetch(path = '', options = {}) {
+    options = {
+      ...options,
+      credentials: 'include',
+    };
+
+    return await fetch(dummy.fileUrl + path, options);
+  }
+
+  #savePersonality() {}
 }
